refactor(analysis): give spatial join its own analysis_sql family module
#1089 names three families — overlay, measure, transform — and spatial join is
in none of them, though it landed in the same batch (#953) as two operations
the issue does place. Rather than stretch a name over it, it takes a fourth
module.

It sits between the two it half-resembles. Like overlay it reads a second
layer, but nothing is cut or filtered — the output geometry is the source
feature verbatim — so it has no mask, no ST_Dimension = 2 guard and nothing to
subdivide, which is every specific thing the issue says overlay members share.
Like measure it renders a (select_columns, join_clause) pair and leaves the
geometry alone, but it is not a measurement, and it is the larger of the two by
three times.

What travels with it is machinery no other family touches: MAX_IDENTIFIER_LENGTH
and the truncation that makes the collision guards compare the names PostgreSQL
will actually create, the join-field identifier rule and its cap, and the
two-lateral count/pick shape with its lowest-gid tie-break. Folding that into
measure.py would put the join-field rules in the path of anyone reading about
area.

Byte-identity: all 94 cases pass.
